+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## v3.1.10 (patch)
+
+Changes since v3.1.9:
+
+- Bump the ktsu group with 9 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+## v3.1.9 (patch)
+
+Changes since v3.1.8:
+
+- Bump the ktsu group with 9 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+## v3.1.8 (patch)
+
+Changes since v3.1.7:
+
+- Bump the ktsu group with 14 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+## v3.1.7 (patch)
+
+Changes since v3.1.6:
+
+- Bump the ktsu group with 6 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+## v3.1.6 (patch)
+
+Changes since v3.1.5:
+
+- Bump the ktsu group with 9 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
+## v3.1.5 (patch)
+
+Changes since v3.1.4:
+
+- Bump the ktsu group with 12 updates ([@dependabot[bot]](https://github.com/dependabot[bot]))
+
 ## v3.1.4 (patch)
 
 Changes since v3.1.3: